// packages/react-wallet-kit/src/types/base.ts
export enum ImportType {
  Wallet = "WALLET",
  PrivateKey = "PRIVATE_KEY",
}